NativeBufferUtils: fixed `memorySet(Buffer, int, long)` parameters

+++ jme3-alloc/src/main/java/com/jme3/alloc/NativeBufferAllocator.java
@@ -15,13 +15,13 @@ public final class NativeBufferAllocator {
     }
 
     /**
-     * Creates a new direct byte buffer explicitly with a specific size.
+     * Creates a new direct byte buffer explicitly with a specific [capacity] in bytes.
      *
-     * @param size the byte buffer size used for allocating the buffer.
-     * @return a new direct byte buffer object.
-     * @see com.jme3.alloc.util.NativeBufferUtils#calloc(int)
+     * @param size the buffer size in bytes units
+     * @return a new direct byte buffer object of size [capacity] in bytes
+     * @see com.jme3.alloc.util.NativeBufferUtils#clearAlloc(long)
      */
-    public static ByteBuffer createDirectByteBuffer(final int capacity) {
+    public static ByteBuffer createDirectByteBuffer(final long capacity) {
         return (ByteBuffer) NativeBufferUtils.clearAlloc(capacity);
     }
     
@@ -29,7 +29,7 @@ public final class NativeBufferAllocator {
      * Releases the memory of a direct buffer using a buffer object reference.
      *
      * @param buffer the buffer reference to release its memory.
-     * @see com.jme3.alloc.util.NativeBufferUtils#free(java.nio.Buffer)
+     * @see com.jme3.alloc.util.NativeBufferUtils#destroy(java.nio.Buffer)
      */
     public static void releaseDirectByteBuffer(final ByteBuffer buffer) {
         NativeBufferUtils.destroy(buffer);
